[+] Вынесение генерации шаблонов в библиотеку
