+++ GPO BLAZOR/GPO BLAZOR/GPO BLAZOR.csproj
@@ -15,21 +15,18 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="PDFConstructor\FormTemplate\**" />
     <Compile Remove="DBAgent\**" />
     <Compile Remove="wwwroot\Новая папка\**" />
-    <Content Remove="PDFConstructor\FormTemplate\**" />
     <Content Remove="DBAgent\**" />
     <Content Remove="wwwroot\Новая папка\**" />
-    <EmbeddedResource Remove="PDFConstructor\FormTemplate\**" />
     <EmbeddedResource Remove="DBAgent\**" />
     <EmbeddedResource Remove="wwwroot\Новая папка\**" />
-    <None Remove="PDFConstructor\FormTemplate\**" />
     <None Remove="DBAgent\**" />
     <None Remove="wwwroot\Новая папка\**" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\PdfFilePrinting\PdfFilePrinting.csproj" />
     <ProjectReference Include="..\GPO BLAZOR.Client\GPO BLAZOR.Client.csproj" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.0" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="9.0.0" />
@@ -44,7 +41,7 @@
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="9.0.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
     <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.21.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="3.6.3" />
+    <PackageReference Include="MSTest.TestFramework" Version="3.6.4" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.1" />
     <PackageReference Include="PDFsharp" Version="6.1.1" />
