--- GPO BLAZOR/GPO BLAZOR/PDFConstructor/FieldModel/FirstNameTextField.xml
@@ -1,28 +1,0 @@
-﻿<?xml version="1.0"?>
-
-<Body DBName="gpo2">
-	<SQLQuery>
-		<get>
-		SELECT *
-		From DBTableName
-		</get>
-		<set>
-			
-		</set>
-		<post>
-			
-		</post>
-		<delete>
-			
-		</delete>
-	</SQLQuery>
-	<PrimaryTable TableName="NamePrimarytable" CollumName="NamePrimaryCollum">
-		
-		<ForeignTable>
-			
-		</ForeignTable>
-		<ForeignTable>
-
-		</ForeignTable>
-	</PrimaryTable>
-</Body>
