--- GPO BLAZOR/GPO BLAZOR/PDFConstructor/SelectedAtributes/Groups.xml
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0"?>
-<Items>
-	<Item>
-		A1
-	</Item>
-	<Item>
-		A2
-	</Item>
-	<Item>
-		A3
-	</Item>
-	<Item>
-		A4
-	</Item>
-</Items>
