--- GPO BLAZOR/GPO BLAZOR/PDFConstructor/TestFile.xml
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0"?>
-<A1>
-	
-</A1>
