--- PdfFilePrinting/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace PdfFilePrinting
-{
-    public class Class1
-    {
-
-    }
-}
